portfolio-site/requirements.txt
langchain==0.3.13
langchain-google-genai==2.0.7
google-generativeai==0.8.4
numpy==2.2.1
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
python-frontmatter==1.1.0